Update deps, improve some PostCSS logging
