+++ lib/compile-stylesheets.js
@@ -17,18 +17,9 @@ const SourceMapConsumer = require('source-map').SourceMapConsumer;
 const constants = require('./constants');
 const postcssAbsoluteUrls = require('./postcss-absolute-urls');
 const joinUrlParts = require('./join-url-parts');
+const handlePostcssError = require('./handle-postcss-error');
 
 const urlCache = new Map();
-
-// Special logging for PostCSS errors. Makes them very easy to understand.
-function handlePostcssError(error) {
-  if (error.name === 'CssSyntaxError') {
-    console.log(error.message);
-    console.log(error.showSourceCode());
-  } else {
-    throw error;
-  }
-}
 
 /**
  * Fetch and concatenate stylesheets, with source maps.
