+++ lib/inline-css.js
@@ -5,6 +5,7 @@ const globby = require('globby');
 const workerFarm = require('worker-farm');
 const timelog = require('./timelog');
 const workers = workerFarm(require.resolve('./css-inlining-worker'));
+const handlePostcssError = require('./handle-postcss-error');
 
 /**
  * Inline CSS in static HTML pages.
@@ -64,8 +65,12 @@ function inlineCss(htmlRoot, cssPath, options) {
     })
     .then(() => {
       if (errors.length) {
-        errors.forEach(e => {
-          console.error(e.message);
+        errors.forEach(error => {
+          if (error.name === 'CssSyntaxError') {
+            handlePostcssError(error);
+          } else {
+            console.error(error.message);
+          }
         });
         throw new Error('Errors while inlining CSS');
       } else {
