+++ package.json
@@ -68,23 +68,23 @@
   "dependencies": {
     "@mapbox/babel-plugin-transform-jsxtreme-markdown": "^0.4.0",
     "@mapbox/jsxtreme-markdown-loader": "^0.5.0",
-    "@mapbox/link-hijacker": "^0.3.1",
+    "@mapbox/link-hijacker": "^0.3.2",
     "@mapbox/link-to-location": "^0.1.0",
-    "@mapbox/postcss-html-filter": "^0.2.0",
+    "@mapbox/postcss-html-filter": "^0.3.1",
     "@mapbox/scroll-restorer": "^0.1.0",
     "assets-webpack-plugin": "^3.5.1",
     "autoprefixer": "^7.1.1",
-    "babel-loader": "^7.0.0",
+    "babel-loader": "^7.1.2",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-plugin-transform-react-jsx-self": "^6.22.0",
     "babel-plugin-transform-react-jsx-source": "^6.22.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.6",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.8",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
-    "babel-register": "^6.24.1",
+    "babel-register": "^6.26.0",
     "browser-sync": "^2.18.12",
-    "chalk": "^2.0.0",
+    "chalk": "^2.1.0",
     "chokidar": "^1.7.0",
     "concat-with-sourcemaps": "^1.0.4",
     "connect-history-api-fallback": "^1.3.0",
@@ -96,20 +96,20 @@
     "del": "^3.0.0",
     "es6-promise": "^4.1.0",
     "file-loader": "^0.11.2",
-    "get-port": "^3.1.0",
+    "get-port": "^3.2.0",
     "globby": "^6.1.0",
     "got": "^7.0.0",
-    "gray-matter": "^3.0.2",
+    "gray-matter": "^3.0.3",
     "hasha": "^3.0.0",
-    "html-webpack-plugin": "^2.28.0",
+    "html-webpack-plugin": "^2.30.1",
     "ignore-loader": "^0.1.2",
     "json-loader": "^0.5.4",
     "lodash": "^4.17.4",
     "meow": "^3.7.0",
     "mkdirp": "^0.5.1",
     "pify": "^3.0.0",
-    "postcss": "^6.0.1",
-    "postcss-url": "^7.1.1",
+    "postcss": "^6.0.9",
+    "postcss-url": "^7.1.2",
     "postcss-values-parser": "^1.2.2",
     "prop-types": "^15.5.10",
     "remark-lint-no-dead-urls": "^0.2.0",
@@ -117,14 +117,14 @@
     "resolve-from": "^3.0.0",
     "sitemap-static": "^0.4.2",
     "slugg": "^1.2.1",
-    "source-map-support": "^0.4.15",
+    "source-map-support": "^0.4.16",
     "tempy": "^0.1.0",
     "uglify-js": "^2.8.29",
-    "webpack": "^3.4.0",
+    "webpack": "^3.5.5",
     "webpack-chunk-hash": "^0.4.0",
     "webpack-merge": "^4.1.0",
     "webpack-parallel-uglify-plugin": "^0.4.2",
-    "worker-farm": "^1.4.1",
+    "worker-farm": "^1.5.0",
     "xtend": "^4.0.1"
   },
   "peerDependencies": {
@@ -134,17 +134,17 @@
   },
   "devDependencies": {
     "@mapbox/vnu-validate-html": "^0.1.0",
-    "babel-cli": "^6.24.1",
+    "babel-cli": "^6.26.0",
     "babel-eslint": "^7.2.3",
     "enzyme": "^2.8.2",
     "enzyme-to-json": "^1.5.1",
-    "eslint": "^4.0.0",
+    "eslint": "^4.5.0",
     "eslint-plugin-filenames": "^1.2.0",
     "eslint-plugin-node": "^5.0.0",
-    "eslint-plugin-react": "^7.0.1",
+    "eslint-plugin-react": "^7.3.0",
     "husky": "^0.14.1",
     "jest": "^20.0.4",
-    "lint-staged": "^4.0.0",
+    "lint-staged": "^4.0.3",
     "path-type": "^3.0.0",
     "prettier": "^1.5.2",
     "react": "^15.5.4",
